+++ NEWS.md
@@ -2,6 +2,8 @@
 
 *  The arguments `skip.vertex`, `skip.edges`, and `skip.arrows` now work as
    documented.
+  
+*  New function `nplot_legend()` helps adding legends to the figure.
 
 
 # netplot 0.1-0